deploy/monitoring/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
evidently
pandas
numpy
